♻️  refactor: error와 exception 분리 #1
