+++ src/main/java/sopt/org/motivooServer/domain/user/exception/UserException.java
@@ -1,0 +1,11 @@
+package sopt.org.motivooServer.domain.user.exception;
+
+import sopt.org.motivooServer.global.common.error.ErrorType;
+import sopt.org.motivooServer.global.common.exception.BusinessException;
+
+public class UserException extends BusinessException {
+
+	public UserException(ErrorType errorType) {
+		super(errorType);
+	}
+}
