+++ src/main/java/sopt/org/motivooServer/global/common/error/ErrorType.java
@@ -1,4 +1,4 @@
-package sopt.org.motivooServer.global.common.exception;
+package sopt.org.motivooServer.global.common.error;
 
 import org.springframework.http.HttpStatus;
 
@@ -8,14 +8,12 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public enum SuccessType {
+public enum ErrorType {
 
 	/**
-	 * 200 Ok
+	 * 500 Internal Server Error
 	 */
-	HEALTH_CHECK_SUCCESS(HttpStatus.OK, "헬스체크용 API 호출에 성공했습니다.")
-
-
+	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "알 수 없는 서버 에러가 발생했습니다"),
 
 	;
